+++ chalice/cli/__init__.py
@@ -20,7 +20,9 @@ from chalice.awsclient import TypedAWSClient
 from chalice.cli.factory import CLIFactory
 from chalice.config import Config  # noqa
 from chalice.logs import display_logs
-from chalice.utils import create_zip_file, record_deployed_values
+from chalice.utils import create_zip_file
+from chalice.utils import record_deployed_values
+from chalice.utils import remove_stage_from_deployed_values
 from chalice.deploy.deployer import validate_python_version
 from chalice.utils import getting_started_prompt
 from chalice.constants import CONFIG_VERSION, TEMPLATE_APP, GITIGNORE
@@ -116,11 +118,30 @@ def deploy(ctx, autogen_policy, profile, api_gateway_stage, stage,
     factory = ctx.obj['factory']  # type: CLIFactory
     factory.profile = profile
     config = factory.create_config_obj(
-        chalice_stage_name=stage, autogen_policy=autogen_policy)
+        chalice_stage_name=stage, autogen_policy=autogen_policy,
+        api_gateway_stage=api_gateway_stage,
+    )
     session = factory.create_botocore_session()
     d = factory.create_default_deployer(session=session, prompter=click)
     deployed_values = d.deploy(config, chalice_stage_name=stage)
     record_deployed_values(deployed_values, os.path.join(
+        config.project_dir, '.chalice', 'deployed.json'))
+
+
+@cli.command('delete')
+@click.option('--profile', help='Override profile at deploy time.')
+@click.option('--stage', default=DEFAULT_STAGE_NAME,
+              help='Name of the Chalice stage to delete.')
+@click.pass_context
+def delete(ctx, profile, stage):
+    # type: (click.Context, str, str) -> None
+    factory = ctx.obj['factory']  # type: CLIFactory
+    factory.profile = profile
+    config = factory.create_config_obj(chalice_stage_name=stage)
+    session = factory.create_botocore_session()
+    d = factory.create_default_deployer(session=session, prompter=click)
+    d.delete(config, chalice_stage_name=stage)
+    remove_stage_from_deployed_values(stage, os.path.join(
         config.project_dir, '.chalice', 'deployed.json'))