+++ chalice/deploy/deployer.py
@@ -3,6 +3,7 @@
 Handles Lambda and API Gateway deployments.
 
 """
+from __future__ import print_function
 import json
 import os
 import sys
@@ -15,7 +16,7 @@ from typing import Any, Tuple, Callable, List, Dict, Optional  # noqa
 from chalice import app  # noqa
 from chalice import __version__ as chalice_version
 from chalice import policy
-from chalice.awsclient import TypedAWSClient
+from chalice.awsclient import TypedAWSClient, ResourceDoesNotExistError
 from chalice.config import Config, DeployedResources  # noqa
 from chalice.deploy.packager import LambdaDeploymentPackager
 from chalice.deploy.swagger import SwaggerGenerator
@@ -151,6 +152,16 @@ class Deployer(object):
         self._apigateway_deploy = apigateway_deploy
         self._lambda_deploy = lambda_deploy
 
+    def delete(self, config, chalice_stage_name=DEFAULT_STAGE_NAME):
+        # type: (Config, str) -> None
+        existing_resources = config.deployed_resources(chalice_stage_name)
+        if existing_resources is None:
+            print('No existing resources found for stage %s.' %
+                  chalice_stage_name)
+            return
+        self._apigateway_deploy.delete(existing_resources)
+        self._lambda_deploy.delete(existing_resources)
+
     def deploy(self, config, chalice_stage_name=DEFAULT_STAGE_NAME):
         # type: (Config, str) -> Dict[str, Any]
         """Deploy chalice application to AWS.
@@ -201,6 +212,23 @@ class LambdaDeployer(object):
         self._osutils = osutils
         self._app_policy = app_policy
 
+    def delete(self, existing_resources):
+        # type: (DeployedResources) -> None
+        handler_name = existing_resources.api_handler_name
+        role_arn = self._get_lambda_role_arn(handler_name)
+        print('Deleting lambda function %s' % handler_name)
+        try:
+            self._aws_client.delete_function(handler_name)
+        except ResourceDoesNotExistError as e:
+            print('No lambda function named %s found.' % e)
+        if role_arn is not None:
+            role_name = role_arn.split('/')[1]
+            if self._prompter.confirm(
+                    'Delete the role %s?' % role_name,
+                    default=False, abort=False):
+                print('Deleting role name %s' % role_name)
+                self._aws_client.delete_role(role_name)
+
     def deploy(self, config, existing_resources, stage_name):
         # type: (Config, OPT_RESOURCES, str) -> Dict[str, str]
         deployed_values = {}
@@ -233,6 +261,14 @@ class LambdaDeployer(object):
                 "you like to continue? " % (lambda_config['Runtime'],
                                             lambda_python_version),
                 default=True, abort=True)
+
+    def _get_lambda_role_arn(self, role_name):
+        # type: (str) -> Optional[str]
+        try:
+            role_arn = self._aws_client.get_role_arn_for_name(role_name)
+            return role_arn
+        except ValueError:
+            return None
 
     def _get_or_create_lambda_role_arn(self, config, role_name):
         # type: (Config, str) -> str
@@ -342,6 +378,15 @@ class APIGatewayDeployer(object):
         # type: (TypedAWSClient) -> None
         self._aws_client = aws_client
 
+    def delete(self, existing_resources):
+        # type: (DeployedResources) -> None
+        rest_api_id = existing_resources.rest_api_id
+        print('Deleting rest API %s' % rest_api_id)
+        try:
+            self._aws_client.delete_rest_api(rest_api_id)
+        except ResourceDoesNotExistError as e:
+            print('No rest API with id %s found.' % e)
+
     def deploy(self, config, existing_resources, lambda_arn):
         # type: (Config, OPT_RESOURCES, str) -> Tuple[str, str, str]
         if existing_resources is not None and \
@@ -351,9 +396,8 @@ class APIGatewayDeployer(object):
             rest_api_id = existing_resources.rest_api_id
             return self._create_resources_for_api(config, rest_api_id,
                                                   lambda_arn)
-        else:
-            print("Initiating first time deployment...")
-            return self._first_time_deploy(config, lambda_arn)
+        print("Initiating first time deployment...")
+        return self._first_time_deploy(config, lambda_arn)
 
     def _first_time_deploy(self, config, lambda_arn):
         # type: (Config, str) -> Tuple[str, str, str]
@@ -457,12 +501,12 @@ class ApplicationPolicyHandler(object):
             self._osutils.get_file_contents(filename, binary=False)
         )
 
-    def record_policy(self, config, policy):
+    def record_policy(self, config, policy_document):
         # type: (Config, Dict[str, Any]) -> None
         policy_file = self._app_policy_file(config)
         self._osutils.set_file_contents(
             policy_file,
-            json.dumps(policy, indent=2, separators=(',', ': ')),
+            json.dumps(policy_document, indent=2, separators=(',', ': ')),
             binary=False
         )