+++ chalice/local.py
@@ -3,13 +3,15 @@
 This is intended only for local development purposes.
 
 """
+from __future__ import print_function
 import functools
 from collections import namedtuple
+
 from six.moves.BaseHTTPServer import HTTPServer
 from six.moves.BaseHTTPServer import BaseHTTPRequestHandler
+from typing import List, Any, Dict, Tuple, Callable  # noqa
 
 from chalice.app import Chalice, CORSConfig  # noqa
-from typing import List, Any, Dict, Tuple, Callable  # noqa
 from chalice.compat import urlparse, parse_qs